refactor(cli): CLI-V2 Phase 8 — extract shared form helpers to internal/cli/prompt
The interactive form helpers (NewForm/Input/RunConfirm/RunForm/FormTheme/
FormKeyMap/PromptGlyph) lived in internal/install but are pure UI wrappers
over charmbracelet/huh + the brand theme, with zero installer coupling.
The jentic (agent) command tree imported internal/install SOLELY for these
— the last soft coupling forcing the data-plane CLI to link operator code.

Move them to a new internal/cli/prompt package (impl/1.1 §1a). install
re-exports them as aliases so its own callers keep working during the
decomposition; the form guard tests now sanction prompt/prompt.go as the
raw-huh home. This is the enabling step for making internal/cli/api
installer-free (06 §1 / 07 §3 airgap).
