+++ cli/internal/install/form_guard_test.go
@@ -17,14 +17,14 @@ import (
 func TestAllFormsUseSharedConstructors(t *testing.T) {
 	root := moduleRoot(t)
 
-	// The sanctioned homes of the raw huh constructors: install/theme.go (the
+	// The sanctioned homes of the raw huh constructors: internal/cli/prompt (the
 	// shared themed helpers) and the binder's dynamic-form generator, whose output
-	// is themed at run time via install.RunForm (impl/6.1 — the one generic
-	// schema-driven form builder; it cannot import install without inverting the
-	// layering, so it builds raw huh fields and the caller applies the theme).
+	// is themed at run time via prompt.RunForm (impl/6.1 — the one generic
+	// schema-driven form builder; it builds raw huh fields and the caller applies
+	// the theme).
 	allowed := map[string]bool{
-		filepath.Join(root, "internal", "install", "theme.go"):      true,
-		filepath.Join(root, "internal", "cli", "binder", "form.go"): true,
+		filepath.Join(root, "internal", "cli", "prompt", "prompt.go"): true,
+		filepath.Join(root, "internal", "cli", "binder", "form.go"):   true,
 	}
 	forbidden := []string{"huh.NewForm(", "huh.NewInput("}
 
@@ -46,7 +46,7 @@ func TestAllFormsUseSharedConstructors(t *testing.T) {
 		for _, tok := range forbidden {
 			if strings.Contains(src, tok) {
 				rel, _ := filepath.Rel(root, path)
-				t.Errorf("%s calls %s directly; use install.NewForm / install.Input instead", rel, tok)
+				t.Errorf("%s calls %s directly; use prompt.NewForm / prompt.Input instead", rel, tok)
 			}
 		}
 		return nil
@@ -64,8 +64,8 @@ func TestAllFormsUseSharedConstructors(t *testing.T) {
 // on the confirm itself.
 func TestNoConfirmRunEscapeHatch(t *testing.T) {
 	root := moduleRoot(t)
-	// theme.go documents and implements the sanctioned helper; exempt it.
-	allowed := filepath.Join(root, "internal", "install", "theme.go")
+	// prompt.go documents and implements the sanctioned helper; exempt it.
+	allowed := filepath.Join(root, "internal", "cli", "prompt", "prompt.go")
 	// Matches huh.NewConfirm()...Run() across lines, with any builder chain in
 	// between, but stops at the first Run( so it can't span unrelated calls.
 	pattern := regexp.MustCompile(`huh\.NewConfirm\((?s:[^;]*?)\.Run\(`)
